pyproject.toml
[tool.poetry]
name = "repl-nix-nflplayofftracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
trafilatura = "^1.12.2"
flask-sqlalchemy = "^3.1.1"
apscheduler = "^3.10.4"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
